+++ Rebus.Tests/Persistence/SqlServer/TestSqlServerSagaStorage.cs
@@ -1,9 +1,10 @@
 ﻿using System;
+using System.Linq;
 using System.Threading.Tasks;
 using NUnit.Framework;
-using Rebus.Exceptions;
 using Rebus.Logging;
 using Rebus.Persistence.SqlServer;
+using Rebus.Sagas;
 
 namespace Rebus.Tests.Persistence.SqlServer
 {
@@ -13,6 +14,7 @@ namespace Rebus.Tests.Persistence.SqlServer
         SqlServerSagaStorage _storage;
         string _dataTableName;
         DbConnectionProvider _connectionProvider;
+        string _indexTableName;
 
         protected override void SetUp()
         {
@@ -20,18 +22,55 @@ namespace Rebus.Tests.Persistence.SqlServer
             _connectionProvider = new DbConnectionProvider(SqlTestHelper.ConnectionString, loggerFactory);
 
             _dataTableName = TestConfig.QueueName("sagas");
-            var indexTableName = TestConfig.QueueName("sagaindex");
+            _indexTableName = TestConfig.QueueName("sagaindex");
 
-            SqlTestHelper.DropTable(indexTableName);
+            SqlTestHelper.DropTable(_indexTableName);
             SqlTestHelper.DropTable(_dataTableName);
 
-            _storage = new SqlServerSagaStorage(_connectionProvider, _dataTableName, indexTableName, loggerFactory);
+            _storage = new SqlServerSagaStorage(_connectionProvider, _dataTableName, _indexTableName, loggerFactory);
         }
 
         [Test]
-        public async Task ThrowsExceptionWhenInitializeOnOldSchema()
+        public async Task DoesNotThrowExceptionWhenInitializeOnOldSchema()
         {
-            var createTableOldSchema = $@"
+            await CreatePreviousSchema();
+
+            _storage.Initialize();
+
+            _storage.EnsureTablesAreCreated();
+        }
+
+        [Test]
+        public async Task CanRoundtripSagaOnOldSchema()
+        {
+            var noProps = Enumerable.Empty<ISagaCorrelationProperty>();
+
+            await CreatePreviousSchema();
+
+            _storage.Initialize();
+
+            var sagaData = new MySagaDizzle {Id=Guid.NewGuid(), Text = "whee!"};
+
+            await _storage.Insert(sagaData, noProps);
+
+            var roundtrippedData = await _storage.Find(typeof(MySagaDizzle), "Id", sagaData.Id.ToString());
+
+            Assert.That(roundtrippedData, Is.TypeOf<MySagaDizzle>());
+            var sagaData2 = (MySagaDizzle)roundtrippedData;
+            Assert.That(sagaData2.Text, Is.EqualTo(sagaData.Text));
+        }
+
+        class MySagaDizzle : ISagaData
+        {
+            public Guid Id { get; set; }
+            public int Revision { get; set; }
+            public string Text { get; set; }
+        }
+
+        async Task CreatePreviousSchema()
+        {
+            var createTableOldSchema =
+                $@"
 
 CREATE TABLE [dbo].[{_dataTableName}](
 	[id] [uniqueidentifier] NOT NULL,
@@ -45,7 +84,23 @@ CREATE TABLE [dbo].[{_dataTableName}](
 
 ";
 
-            Console.WriteLine($"Creating table {_dataTableName}");
+            var createTableOldSchema2 =
+                $@"
+
+CREATE TABLE [dbo].[{_indexTableName}](
+	[saga_type] [nvarchar](40) NOT NULL,
+	[key] [nvarchar](200) NOT NULL,
+	[value] [nvarchar](200) NOT NULL,
+	[saga_id] [uniqueidentifier] NOT NULL,
+ CONSTRAINT [PK_{_indexTableName}] PRIMARY KEY CLUSTERED 
+ (
+	[key] ASC,
+	[value] ASC,
+	[saga_type] ASC
+ ))
+";
+
+            Console.WriteLine($"Creating tables {_dataTableName} and {_indexTableName}");
 
             using (var connection = await _connectionProvider.GetConnection())
             {
@@ -54,18 +109,14 @@ CREATE TABLE [dbo].[{_dataTableName}](
                     command.CommandText = createTableOldSchema;
                     command.ExecuteNonQuery();
                 }
+                using (var command = connection.CreateCommand())
+                {
+                    command.CommandText = createTableOldSchema2;
+                    command.ExecuteNonQuery();
+                }
 
                 await connection.Complete();
             }
-
-            Console.WriteLine("Telling saga storage to create its schema");
-
-            var exception = Assert.Throws<AggregateException>(() =>
-            {
-                _storage.EnsureTablesAreCreated();
-            });
-
-            Console.WriteLine(exception);
         }
     }
 }